making the autocommands way better
